refactor: watcher_count DDL 수정 및 Java 필드 타입 변경

+++ src/main/java/com/codeit/team5/mopl/content/entity/ContentStats.java
@@ -35,7 +35,7 @@ public class ContentStats {
     private double ratingSum;
 
     @Column(nullable = false)
-    private int watcherCount;
+    private long watcherCount;
 
     @Column(nullable = false)
     private Instant updatedAt;
